cleanup, rename test coordinator
