+++ gradle.properties
@@ -1,4 +1,4 @@
-canaryVersion=1.0
+canaryVersion=1.1
 
 minestomVariant=Minestom
 minestomVersion=ebb179da7c
